moved test to tests/
