+++ verifier/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-#tlsn-mpc-core = { path = "../mpc/mpc-core", features = ["garble"] }
 blake3 = "1.3.3"
 sha2 = "0.10"
 p256 = { version = "0.10", features = ["ecdsa"]}
@@ -22,4 +21,7 @@ serde = { version = "1.0", features = ["derive"] }
 bincode = "1.3.3"
 
 [dev-dependencies]
+tlsn-mpc-core = { path = "../mpc/mpc-core" }
+tlsn-mpc-circuits = { path = "../mpc/mpc-circuits" }
+tlsn-tls-circuits = { path = "../tls/tls-circuits" }
 hex = "0.4"
